Add files via upload
The CW2PuTorchTensor takes a message and converts it to a numpy array for given tone and code speed setting and then saves the array as a PyTorch tensor.  This is version 0.001 of the code so likely needs work.  A single letter tensor was approximately 300 kb and messages can take a lot of memory.  On a windows Anaconda Spyder the plot worked well in console mode but on the authors machine froze if automatic window plot was chosen.  Filesnames are not checked and will over write a previous save.  Fortunately you can create a bunch of tensor files of interest quickly.