Add files via upload
Real morse code captured during a QSO contest on the 40 meter band Sunday afternoon - the spectrogram was created with Audacity.  I have been looking for a python code version to do the same thing but so far not satisfied with my efforts so will be a later post.  Audacity worked pretty well for this visualization but likely not fast enough for real-time with a deep learning modulator to translate the patterns to english.  Note the audio here has about a 12khz bandwidth to show multiple tones heard.  I was using a SDRplay radio with the digital window.  The band width was limited by the usb sound capture dongle that was used to about 12 khz.  Audio sampling rate was 44100.