Add files via upload
CW_spectrogram version 1 is based on spectrum analyzer codes and a blog article by SBarratt and Yumi as well as some clippings from the matplotlib manual.  The code opens a microphone audio channel and allows realtime monitoring of the incoming signal as an FFT display.  The signal is accumulated and plotted as an image.  The captured code from a  receiver tuned to 40m band.   A user might need to adjust the display ranges is long captures as the autoranging needs some work.  The twight color map seemed to work okay for the morse code display.  An alternative spectrogram and waveform is also displayed.  The Blackman filter greatly improved the FFT for morse code display so  is included.  This is a first version of this and error handling of exiting sometimes failed to close. Suggest run in a deverlopers environment.