Add files via upload
Add a second version of the CW_spectrpgram which runs on the raspberry pi 4.  The pyaudio version had issues using my usb sound dongle and so rewrote to use the sounddevice module instead of the pyaudio.  The program still uses pyqtgraph for audio tones display.  The actual spectrogram is plotted using matplotlib and at this time you need to use the pan feature to zoom the display to see code details.  Example image is also being uploaded