Add files via upload
Creates a morse code pattern from assembly of numpy arrays and then performs a FFT on the wavefile to visualize the code pattern in the frequency domain.  Data acquisition allows the input audio to be viewed as a frequency in real time.  The resulting spectrogram is viewed as a displayed image with axis scaling.  There is a bug on reading back the wave file but the wave files play well in windows media player.